requirements.txt
img2vec_pytorch==1.0.2
scikit-learn==1.7.2
pillow==12.0.0